+++ internal/cluster/icluster/cluster.go
@@ -1,9 +1,9 @@
 package icluster
 
 import (
-	"context"
-
 	"google.golang.org/grpc"
+
+	"github.com/ydb-platform/ydb-go-sdk/v3/internal/closer"
 )
 
 type Cluster interface {
@@ -12,7 +12,5 @@ type Cluster interface {
 	// Lazy getting grpc-connection must use for embedded client-side balancing
 	// DB may be put into code-generated client constructor as is.
 	grpc.ClientConnInterface
-
-	// Close clears resources and close all connections to YDB
-	Close(ctx context.Context) error
+	closer.Closer
 }