+++ internal/errors/retryable.go
@@ -1,0 +1,11 @@
+package errors
+
+type RetryableError struct {
+	Msg               string
+	BackoffType       BackoffType
+	MustDeleteSession bool
+}
+
+func (e *RetryableError) Error() string {
+	return e.Msg
+}