+++ CHANGELOG.md
@@ -1,10 +1,10 @@
 ## 3.7.0
-* Replaced `Option` to `CustomOption` on `Connection` interface methods 
+* Replaced `Option` to `CustomOption` on `Connection` interface methods
 * Implements `WithCustom[Token,Database]` options for redefine database and token
 * Removed experimental `balancer.PreferEndpoints[WithFallback][RegEx]` balancers
-* Implements `balancer.PreferLocations[WithFallback]` balancers
 * Supported connections `TTL` with `Option` `WithConnectionTTL`
-* Remove unnecessary `WithFastDial` option (lazy connections always fast inserts into cluster)
+* Remove unnecessary `WithFastDial` option (lazy connections are always fast inserts into cluster)
+* Added `ScriptingYQL` service client
 
 ## 3.6.2
 * Refactored table retry helpers